Moved encode extension method from Encodable and make a part of FileManager
