+++ Sources/Extensions/Foundation/DecodableExtensions.swift
@@ -1,0 +1,8 @@
+//
+// DecodableExtensions.swift
+// SwifterSwift
+//
+// Created by Alexei Kozachenko on 2018-10-29.
+// Copyright © 2018 SwifterSwift. All rights reserved.
+
+import Foundation
